+++ src/linklayer/shaper/AS6802/CoRE4INET_RCShaper.h
@@ -109,7 +109,6 @@ class RCShaper : public TC, public virtual Timed
          * one. Else it saves the state and sends the message immediately when it is
          * received.
          *
-         * @param msg the message to be queued
          */
         virtual void requestPacket();
 
@@ -163,7 +162,7 @@ void RCShaper<TC>::initialize(int stage)
     {
         Timed::initialize();
 
-        numRcPriority = par("numRCpriority").longValue();
+        numRcPriority = (size_t)par("numRCpriority").longValue();
         for (unsigned int i = 0; i < numRcPriority; i++)
         {
             char strBuf[32];
@@ -245,8 +244,8 @@ void RCShaper<TC>::enqueueMessage(cMessage *msg)
         int priority = msg->getSenderModule()->par("priority").longValue();
         if (priority > 0 && (unsigned int) priority < numRcPriority)
         {
-            rcQueue[priority].insert(msg);
-            cComponent::emit(rcQueueLengthSignals[priority], (unsigned long) rcQueue[priority].length());
+            rcQueue[(size_t)priority].insert(msg);
+            cComponent::emit(rcQueueLengthSignals[priority], (unsigned long) rcQueue[(size_t)priority].length());
             TC::notifyListeners();
         }
         else
